backend/requirements.txt
fastapi
langchain
langchain-community
langchain-core
raganything
python-dotenv
langchain_openai